refactor: tiptap용 커스텀 youtubeExtension 최적화 작업
